refactor(settings): extract ConnectorsPanel for Slack integration page
Split the connector list, empty state, and modals out of ConnectorsManager
into a headerless ConnectorsPanel so the Slack integration settings screen
can reuse them under its own SettingsSection without duplicating the
"Slack bot connectors" heading and description. ConnectorsManager now
wraps ConnectorsPanel with its standalone header for the connectors page
and desktop settings dialog.
